completamento back-end dettaglio squadre

+++ Java/quadrangolare_calcio/src/main/java/com/example/quadrangolare_calcio/dao/SquadraDao.java
@@ -3,5 +3,5 @@ package com.example.quadrangolare_calcio.dao;
 import com.example.quadrangolare_calcio.model.Squadra;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SquadraDao extends CrudRepository<Squadra, Integer> {
+public interface SquadraDao extends CrudRepository<Squadra, Long> {
 }
